56. Merge Intervals
再度複習區間是否重疊, 需要比較新進來的interval 最小值, 是否比原interval最大值還要小
這樣的話就可以確定是重疊

Intuition:-
After reading the question we get that we need to merge the over lapping intervals and return the result by keeping the rest of the elements as it is.
To find the overlapping intervals the intervals should be atleast sorted think about this .
Thus for solving we need to sort and then check the consecutive intervals. Once we find the overlapping interval. we will take the max element from it.

Algorithm:-

Firstly, the base case : if there are no intervals return [] .
Sort the intervals .
While traversing the intervals vector we will come accross two coditions
First condition : if there is a overlapping between the intervals then just take out the max element from the ending point and thus we merged them
eg:- [1,4],[2,8] =Mergerd intervals will be> [1,8]
second condition : if there is no overlapping then simply push those interval to our resultant vector .
Now let's move our finger's to coding

Time complexity : O(nlogn) [for using the sort function] + O(n) [for traversing the interval] = O(nlogn)
Space Complexity : O(n) [For our resultant vector]